requirements.txt
# AI Research Agent - Comprehensive Dependencies
# Install with: pip install -r requirements.txt

# ===== CORE WEB FRAMEWORK =====
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0

# ===== AI & LANGUAGE MODELS =====
openai>=1.14.3
sentence-transformers
faiss-cpu
tiktoken>=0.5.0

# ===== BROWSER AUTOMATION =====
# Selenium (for DocSend processing with stealth capabilities)
selenium>=4.15.0
webdriver-manager>=4.0.0

# Playwright (for advanced web scraping)
playwright>=1.50.0

# ===== OCR & IMAGE PROCESSING =====
pytesseract>=0.3.10
Pillow>=10.0.0

# ===== DOCUMENT PROCESSING =====
PyMuPDF  # For PDF processing (fitz)
python-docx>=1.1.0  # For DOCX processing
PyPDF2>=3.0.1  # Alternative PDF processing

# ===== HTTP CLIENTS & WEB =====
aiohttp>=3.9.0
httpx>=0.27.0  # For async HTTP requests
requests>=2.31.0
certifi>=2023.11.17  # For SSL certificate verification
beautifulsoup4
validators
brotli>=1.1.0  # For Brotli compression support in sitemaps

# ===== DATA PROCESSING =====
pandas>=2.1.0
numpy>=1.24.0

# ===== MCP & CRYPTO ANALYSIS =====
# MCP (Model Context Protocol) Dependencies
aiofiles>=23.2.1  # Already included above
asyncio>=3.4.3
websockets>=12.0  # For MCP SSE connections
pydantic>=2.0.0  # For data validation and models

# Visualization & Charting
plotly>=5.17.0  # Interactive charts
altair>=5.0.0   # Statistical visualizations  
bokeh>=3.0.0    # Alternative plotting library

# ===== AUTHENTICATION & SECURITY =====
bcrypt>=4.1.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0  # JWT encoding/decoding

# ===== BACKGROUND JOBS =====
arq>=0.26.0  # Redis-backed async job queue

# ===== FILE VALIDATION =====
python-magic>=0.4.27  # MIME type detection (requires libmagic system library)

# ===== CONFIGURATION & ENVIRONMENT =====
python-dotenv>=1.0.1
pyyaml>=6.0.1

# ===== CACHING & STORAGE =====
redis

# ===== NOTION INTEGRATION =====
notion-client>=2.2.1

# ===== FIRECRAWL & WEB SCRAPING =====
firecrawl-py>=2.4.0

# ===== ASYNC UTILITIES =====
aiofiles>=23.2.1
asyncio-throttle>=1.0.2
tenacity>=8.0.0

# ===== DATE & TIME =====
python-dateutil>=2.9.0

# ===== CLI & UI UTILITIES =====
prompt-toolkit>=3.0.0
rich>=13.7.0
typer>=0.9.0

# ===== SEARCH & RESEARCH =====
duckduckgo-search==7.5.0
google-api-python-client

# ===== OPEN DEEP RESEARCH (ODR) INTEGRATION =====
langgraph>=0.2.55
langchain-community>=0.3.9
langchain-openai>=0.3.7
langchain-anthropic>=0.3.15
langchain-mcp-adapters>=0.1.6
langchain-deepseek>=0.1.2
langchain-tavily
langchain-groq>=0.2.4
tavily-python>=0.5.0
arxiv>=2.1.3
xmltodict>=0.14.2
linkup-sdk>=0.2.3
exa-py>=1.8.8
markdownify>=0.11.6
azure-identity>=1.21.0
azure-search>=1.0.0b2
azure-search-documents>=11.5.2
langsmith>=0.3.37
langchain-google-vertexai>=2.0.25
langchain-google-genai>=2.1.5
supabase>=2.15.3
mcp>=1.9.4

# ===== SPECIALIZED LIBRARIES =====
kitchenai-whisk

# ===== ENTITY EXTRACTION =====
# Pin to specific version to avoid API drift (prompt_description vs instructions, etc.)
langextract==0.1.0

# ===== TESTING =====
pytest>=7.4.0

# ===== SYSTEM DEPENDENCIES NOTES =====
#
# The following system dependencies are required:
#
# BROWSERS (at least one required):
# - Chrome/Chromium (recommended for DocSend)
# - Firefox
# - Microsoft Edge
#
# OCR ENGINE:
# - macOS: brew install tesseract
# - Windows: Download from https://github.com/UB-Mannheim/tesseract/wiki
# - Linux: sudo apt-get install tesseract-ocr (Ubuntu/Debian)
#          sudo yum install tesseract (CentOS/RHEL)
#
# LIBMAGIC (required for langextract entity extraction):
# - macOS: brew install libmagic
# - Windows: pip install python-magic-bin (includes bundled libmagic)
# - Linux: sudo apt-get install libmagic1 (Ubuntu/Debian)
#          sudo yum install file-libs (CentOS/RHEL)
#
# PLAYWRIGHT BROWSERS (for web scraping):
# After pip install, run: playwright install
#
# SUPPORTED PLATFORMS:
# - macOS (Intel/Apple Silicon)
# - Windows 10/11
# - Linux (Ubuntu, CentOS, etc.)
#
# ===== QUICK START =====
#
# 1. Create virtual environment:
#    python3 -m venv venv && source venv/bin/activate
#
# 2. Install system dependencies (macOS example):
#    brew install tesseract libmagic
#
# 3. Install Python packages:
#    pip install -r requirements.txt
#
# 4. Install Playwright browsers:
#    playwright install
#
# 5. Copy .env.example to .env and configure API keys:
#    cp .env.example .env
#
# 6. Run the app:
#    streamlit run main.py
